backend/pyproject.toml
[project]
name = "langchain-rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "streamlit==1.51.0",
    "langchain>=1.0.3",
    "langchain-google-genai>=3.0.0",
    "langgraph>=1.0.2",
    "leann>=0.3.4",
    "python-dotenv>=1.2.1",
    "langchain-core>=1.0.2",
    "pyyaml>=6.0",
    "langchain-openai>=1.0.2",
]
